move concatenate test to separate file
